third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-bottom-up-times.html

<html>
<head>
  <script src="../../http/tests/inspector/inspector-test.js"></script>
  <script src="profiler-test.js"></script>
<script>

function test()
{
    var profileAndExpectations = {
        "title": "profile1",
        "target": function() {
          return WebInspector.targetManager.targets()[0];
        },
        "_profile": {
            "head": {
                "functionName": "(root)",
                "scriptId": "0",
                "url": "a.js",
                "lineNumber": 0,
                "hitCount": 50,
                "callUID": 1000,
                "children": [
                {
                    "functionName": "(idle)",
                    "scriptId": "0",
                    "url": "a.js",
                    "lineNumber": 1,
                    "hitCount": 300,
                    "callUID": 2,
                    "children": []
                },
                {
                    "functionName": "A",
                    "scriptId": "0",
                    "url": "a.js",
                    "lineNumber": 4642,
                    "hitCount": 250,
                    "callUID": 1001,
                    "children": [
                    {
                        "functionName": "C",
                        "scriptId": "0",
                        "url": "a.js",
                        "lineNumber": 525,
                        "hitCount": 100,
                        "callUID": 2000,
                        "children": [
                        {
                            "functionName": "D",
                            "scriptId": "0",
                            "url": "a.js",
                            "lineNumber": 425,
                            "hitCount": 20,
                            "callUID": 3000,
                            "children": []
                        }
                        ]
                    }
                    ]
                },
                {
                    "functionName": "B",
                    "scriptId": "0",
                    "url": "a.js",
                    "lineNumber": 4662,
                    "hitCount": 150,
                    "callUID": 1002,
                    "children": [
                    {
                        "functionName": "C",
                        "scriptId": "0",
                        "url": "a.js",
                        "lineNumber": 525,
                        "hitCount": 100,
                        "callUID": 2000,
                        "children": [
                        {
                            "functionName": "D",
                            "scriptId": "0",
                            "url": "a.js",
                            "lineNumber": 425,
                            "hitCount": 30,
                            "callUID": 3000,
                            "children": []
                        }
                        ]
                    }
                    ]
                }
                ]
            },
            "idleTime": 0.300,
            "startTime": 0,
            "endTime": 1.000
        }
    };
    var view = new WebInspector.CPUProfileView(profileAndExpectations);
    view.viewSelectComboBox.setSelectedIndex(1);
    view._changeView();
    var tree = view.profileDataGridTree;
    if (!tree)
         InspectorTest.addResult("no tree");
    var node = tree.children[0];
    if (!node)
         InspectorTest.addResult("no node");
    while (node) {
        var url = node.element().children[2].firstChild.textContent;
        InspectorTest.addResult(node.callUID + ": " + node.functionName + " " + node.selfTime + " " + node.totalTime + " " + url + " " + JSON.stringify(node.data));
        node = node.traverseNextNode(true, null, true);
    }
    InspectorTest.completeProfilerTest();
}

</script>
</head>
<body onload="runTest()">
<p>
Tests bottom-up view self and total time calculation in CPU profiler.
</p>
</body>
</html>
