third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-bottom-up-times-expected.txt

Tests bottom-up view self and total time calculation in CPU profiler.

2: (idle) 300 300 (idle) {"function":"(idle)","self-percent":"30.00 %","self":"300.0 ms","total-percent":"30.00 %","total":"300.0 ms"}
1001: A 250 370 A {"function":"A","self-percent":"25.00 %","self":"250.0 ms","total-percent":"37.00 %","total":"370.0 ms"}
2000: C 200 250 C {"function":"C","self-percent":"20.00 %","self":"200.0 ms","total-percent":"25.00 %","total":"250.0 ms"}
1002: B 150 280 B {"function":"B","self-percent":"15.00 %","self":"150.0 ms","total-percent":"28.00 %","total":"280.0 ms"}
3000: D 50 50 D {"function":"D","self-percent":"5.00 %","self":"50.0 ms","total-percent":"5.00 %","total":"50.0 ms"}

Profiler was disabled.
